feat: adiciona filtros em categorias e leilões + segurança de perfil

+++ backend/src/main/java/com/github/larybino/leilao/service/CategoryService.java
@@ -5,14 +5,18 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.github.larybino.leilao.exception.BusinessException;
 import com.github.larybino.leilao.exception.NotFoundException;
 import com.github.larybino.leilao.model.Category;
+import com.github.larybino.leilao.repository.AuctionRepository;
 import com.github.larybino.leilao.repository.CategoryRepository;
 
 @Service
 public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
+    @Autowired
+    private AuctionRepository auctionRepository;
 
     public Category create(Category category) {
         return categoryRepository.save(category);
@@ -32,10 +36,19 @@ public class CategoryService {
 
     public void delete(Long id) {
         Category existingCategory = findById(id);
+
+        long auctionCount = auctionRepository.countByCategoryId(id);
+        if (auctionCount > 0) {
+            throw new BusinessException("Não é possível excluir a categoria, pois ela está vinculada a " + auctionCount + " leilão(ões).");
+        }
         categoryRepository.delete(existingCategory);
     }
 
-    public Page<Category> findAll(Pageable pageable) {
-        return categoryRepository.findAll(pageable);
+    public Page<Category> findAll(String searchTerm, Pageable pageable) {
+        if (searchTerm != null && !searchTerm.trim().isEmpty()) {
+            return categoryRepository.findByName(searchTerm, pageable);
+        } else {
+            return categoryRepository.findAll(pageable);
+        }
     }
 }
